refactor: remove async calls from `usePlaceOrder` hook
